move out common code from v1/collections.py
