+++ backend/portal/api/curation/v1/curation/collections/collection_id/datasets/upload_s3.py
@@ -8,7 +8,7 @@ from backend.api_server.db import dbconnect
 from backend.common.corpora_config import CorporaConfig
 from backend.common.corpora_orm import CollectionVisibility
 from backend.portal.api.app.v1.authorization import owner_or_allowed, is_super_curator
-from backend.portal.api.app.v1.common import get_collection_else_forbidden
+from backend.portal.api.collections_common import get_collection_else_forbidden
 
 sts_client = boto3.client("sts")
 logger = logging.getLogger(__name__)
